# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Immobilière",
        "description": "API pour la gestion des biens immobiliers",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connecter un utilisateur (locataire ou comptable)",
                "description": "Authentifie un utilisateur avec son code_id et mot de passe. Le système vérifie s'il s'agit d'un locataire ou d'un comptable et renvoie un token Sanctum en cas de succès.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Identifiants de l'utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_id",
                                    "password"
                                ],
                                "properties": {
                                    "code_id": {
                                        "type": "string",
                                        "example": "LOC12345"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "description": "Données de l'utilisateur",
                                            "type": "object"
                                        },
                                        "token": {
                                            "description": "Token d'authentification Sanctum",
                                            "type": "string"
                                        },
                                        "user_type": {
                                            "type": "string",
                                            "example": "locataire"
                                        },
                                        "redirect": {
                                            "type": "string",
                                            "example": "http://localhost/locataire/dashboard"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants incorrects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Identifiants incorrects"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Compte désactivé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Compte désactivé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/password/forgot": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Demander la réinitialisation du mot de passe",
                "description": "Envoie un lien de réinitialisation de mot de passe à l'email associé au code_id fourni.",
                "operationId": "sendPasswordResetLink",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_id"
                                ],
                                "properties": {
                                    "code_id": {
                                        "type": "string",
                                        "example": "COMPTA001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien de réinitialisation envoyé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Un lien de réinitialisation a été envoyé à votre email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Aucun email associé"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Réinitialiser le mot de passe avec un token",
                "description": "Définit un nouveau mot de passe pour un utilisateur en utilisant le token de réinitialisation.",
                "operationId": "resetUserPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "code_id",
                                    "type",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "longrandomtokenstring..."
                                    },
                                    "code_id": {
                                        "type": "string",
                                        "example": "LOC12345"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "locataire",
                                            "comptable"
                                        ],
                                        "example": "locataire"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NouveauMotDePasse123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NouveauMotDePasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe réinitialisé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mot de passe réinitialisé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token invalide ou expiré"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation (ex: mots de passe non identiques)"
                    }
                }
            }
        },
...
    },
...
}